Fix magpie directory, add feature, rename variable
Because:
- Pulling the initial commit of the magpie implementation produces a
faulty directory
- The 'instructions' wrapping view needs to display different images
based on the scenario
- The value of 'coin' determines the scenario and could
thus use a more meaningful name

Something went wrong with Franziska's commit of her magpie
implementation, resulting in a faulty directory upon pulling from the
GitHub repo. The magpie directory contains a '.git' directory of its
own when I download it from Google Drive, which I think might be the
cause of the problem. Hence, I delete said '.git' directory.

The 'instructions' wrapping view now correctly displays the image
matching the scenario: 'Pip_dolly.png' for the 'move' scenario and
'Pip_nodolly.png for the 'inspect' scenario. To do this I take
Franziska's 'pic' variable, to which she assigns the filepath of the
correct image, and use string concatenation to form the filepath of the
'instructions' image with the variable in it.

The variable 'coin' which Franziska declares in '02_custom_functions.js'
cleverly solves the problem of determining the scenario (I would
probably have tried to put this information into the trial data somehow,
which would have been way more complicated) but needs a better name. I
simply renamed it to 'scenario', since that is what it determines.