Change text in wrapping views, add text to trials
Because:
- The introduction contains part of the instructions, which is counter
to a clear functional separation of the two corresponding wrapping views
- The instructions omit that the participant is supposed to click on the
pictures in the trials.
- The QUD doesn't display the way it is now
- It could be beneficial if each trial restated what the participant is
supposed to do

I rephrased the introduction and the instructions to better separate
the two and to include that the participant needs to click on the
images to progress in the trials. I removed the bit where we ask the
participant to focus for two minutes, but I think that we should put it
back in. I'm just not sure where it belongs yet. I also renamed the
'talk' view to 'instructions2' for clarity's sake.

The QUD didn't display when the line was put "into the trial data but
'outside' of the individual trials" like in Franziska's version  and
putting it inside the view itself, like I tried in my previous commit,
didn't work either. I ended up pasting the line directly into the data
for each trial and then made some changes to the text itself. Where it
used say, e.g., "The boxes were tall", it now says "This is what Pip
tells Jim: 'The boxes were tall' Which boxes is Pip talking about?". I
think that restating the question like that looks nice (not a good
reason in itself, I know) and reminds participants of their task in case
they forgot.